star-image: Transition between two icons
This replaces the custom drawing of a star by partially drawing a
starred and a non-starred icon. This not only makes the code simpler,
but it follows the GTK 4 recommendations of replacing custom drawings by
widgets and CSS. This fixes deprecations in GTK 4.10 as
gtk_snapshot_render_frame() is going away.

To better handle having different sizes, this adds the pixel-size
property.

This also allows us to more easily use more complex and different
shapes, by using one icon for the starred part and another for the
non-starred part.