update-monitor: stop notifying about unprepared critical updates
GNOME Software is supposed to notify about available updates if:

 * The setting to automatically download updates is disabled, or
 * The update is prepared

In Software 40, this behavior regressed and Software began improperly
notifying about updates before they were prepared. That was mostly fixed
in 64585205437371c459d5de792d50af95e23ba346, except for critical
updates. When an update is critical, Software notifies even if automatic
downloads are enabled and the update is not yet prepared. This is wrong.

If should_download is true and all_downloaded is false, we should never
display a notification, regardless of update criticality.

https://pagure.io/fedora-workstation/issue/107