updates-page: Tweak "Check for Updates" button tooltips
The "Check for Updates button" uses two tooltips: "Refresh" and "Stop Refreshing".
In translations, this can be problematic. "Refresh" means "atualizar" in pt-BR and "atualizar" means "update" in english.
However, this button doesn't initiate the update procedure, it only checks the availability of updates.

This commit changes the tooltip to "Check for Updates" and "Stop" to avoid ambiguity.

Fixes: https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2018