lib: Add GsJobManager for tracking pending jobs
This will provide a central place to query which jobs are currently
ongoing, and which ongoing jobs affect a particular app (for example,
one being displayed in the UI).

This will eventually replace the state/pending-state/progress/etc.
properties on `GsApp`, as they cannot represent the state of more than
one ongoing operation, and can’t give specific details for each type
of operation.

For the moment, though, just introduce the new manager class. It
currently works with `GsPluginJobUpdateApps`, and will be updated to
work with more `GsPluginJob` subclasses in future.

It doesn’t work with the `GS_PLUGIN_ACTION_*`-style jobs, as they are
being phased out.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472