gs-page: Factor out user interaction in GsPluginJobUpdateApps
`GsPage` had some hacky fwupd-specific code in it to prompt the user to
take action on a hardware device as part of the update process. For
example, to unplug and replug a hardware device when its firmware was
being updated.

This was a layering violation. With the `GsPluginJobUpdateApps` API, we
can tidy this up by abstracting it through a signal.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472