Do not scroll top in the search changed handler
gs_search_page_set_text already reloads the page
(and scrolls to the top) if the text really did change.
This fixes a bug where we sometimes sroll to the top after
exiting the details page.
Also some signal blocking functions have been removed,
as they are not working anymore.