gs-update-monitor: Change the last check timestamp only if updates could be received
The "check-timestamp" should be updated only if both "refresh-cache" and
"get-updates" succeeded, instead of immediately after "refresh-cache" call
succeeded, thus when the later fails, it's checked an hour later for the updates
again. Otherwise, for example when the first check after login always fails,
like due to the server is connected later after login, the user never sees
new updates.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1316