gs-summary-tile: Show 3-dot icon while icons download
GsApp now provides enough information for us to show a placeholder
icon while the actual app icon is still downloading. In theory we
have enough granularity to show different icons for more states,
such as "waiting in the download queue" or "downloading", but for
simplicity, for now, just use the same 3-dot icon for all these
states.