gs-plugin-icons: Read interactive flag only once
When GsPluginIcons used a worker thread, we couldn't rely on the
plugin having an up-to-date interactive flag, like we were doing.
Without the worker thread, however, we can query and use the
interactive flag since we're running locally.

Do that. Pass the interactive flag to the refine function, as an
extra parameter.