gs-summary-tile: Don't update icon unnecessarily
The refresh method of GsAppTile, which is the parent class of
GsSummary tile, is called in an idle callback, that is queued after
every and any property of the underlying GsApp changes.

This means that every time GsSummaryTile refreshes, it sets the
GIcon of the app. That would be fine, since GsApp often caches its
icons, but GtkImage doesn't check if the icon has changed, so we
end up invalidating the GtkImage anyway every time that is called.

This is unnecessary, and we can track a little closer if we need to
update the app icon.

Track the current app icons state, and only update icons if the app
icons state has changed. Reset the app icons state if the 'app'
property of the tile has changed.