eos-updater: Drop thread-synchronisation GCond
Previously, worker threads needed to be synchronised with the main
thread in the eos-updater plugin, as the main thread received state
change notifications from the eos-updater daemon which the workers
needed to know about (in a thread safe way).

Now that downloading an update has been ported to use the new
`GsPluginJob` API, it is run in the main thread. It was the last worker
thread which needed to synchronise on state, so the `GCond` can be
dropped.

There are still worker threads in eos-updater (they haven’t all been
ported yet), but the rest of them don’t need to synchronise on state.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472