flatpak: Port from gs_plugin_update() to GsPlugin.update_apps_async()
Most of this commit is simple inlining of `gs_plugin_flatpak_update()`,
which has not changed significantly (just renaming variables and
changing error handling to use `GTask`).

There is now some duplication between the ‘download’ and ‘apply’ phases
of the code in `update_apps_thread_cb()`. That can be tidied up in a
subsequent commit.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472