profiler: Drop unscoped profiling macros
While they worked fine, it was quite easy to accidentally put an early
return between the begin and end macros, and cause a leak.

There seem to be no downsides to always using the scoped versions of the
macros, so drop the unscoped ones.

If this works out well, we can drop the `_SCOPED` suffix from the macro
naming in future.

See https://gitlab.gnome.org/GNOME/gnome-software/-/merge_requests/1617#note_1664076

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>