packagekit: Inline gs_systemd_call_*() helper functions
These functions are each only used in one place, so it simplifies the
code to inline them. It will also simplify upcoming changes.

Note that this *is* a behaviour change for the interactivity flags in
`gs_plugin_packagekit_update_apps_async()`: previously the interactivity
for the `pk_offline_trigger_with_flags()` call there was queried from
`GS_PLUGIN_FLAGS_INTERACTIVE`. Now, it can be queried on a per-operation
basis from the `GsPluginUpdateAppsFlags`.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>