packagekit: Make applying app updates asynchronous
The previous hacky port of the implementation was synchronous, and thus
could block the main thread.

Make it asynchronous so it doesn’t block; that’s fairly straightforward
because it’s a single D-Bus call.

PackageKit doesn’t currently provide an async version of the function we
need, so for the moment I’ve copied the PackageKit implementation inline
and made it async. We can use the PackageKit async method when it’s
added.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472