fwupd: Make most of update_apps_async() be async
Port the `gs_plugin_fwupd_download()` function to be async, and then
port most of `gs_plugin_fwupd_update_apps_async()` to be async to make
use of it.

The only remaining bit which needs to be made async is
`gs_plugin_fwupd_install()`, which will happen in the next commit.

This introduces some functional changes: instead of firmware updates
being each downloaded then applied in sequence, they are now all
downloaded in parallel, and each one is applied when its download is
complete.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472