fwupd: Port gs_plugin_fwupd_install() to be async
This completes the port of `update_apps_async()` in the fwupd plugin:
now it’s entirely async.

There are some problems with the old-style progress reporting using
`gs_app_set_progress()` via `self->app_current`: the fwupd API only
provides progress information for an unspecified current device, and
cannot support progress reporting for multiple in-flight operations, so
we cannot reflect that correctly in the progress state of the `GsApp`s.

See https://github.com/fwupd/fwupd/issues/5522

We’ll have to try and improve on this in future.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472