packagekit: Remove unnecessary type check
The type is already checked with the instance cast above, and we can be
pretty sure it’s going to be correct due to the way that
`gs_plugin_packagekit_auto_prepare_update_thread()` is called.

`g_return_if_fail()` is more typically reserved for checking calls into
functions from third-party code, which this is not.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>