gs-app-permissions: Add GS_APP_PERMISSIONS_FLAGS_SYSTEM_DEVICES
And use it in flatpak.

This is distinct from `FLAGS_DEVICES` in that it represents system
devices such as `/dev/shm` which aren’t physical devices such as webcams
or gaming controllers.

As such, they need to be presented in the UI slightly differently,
because the user is less likely to be able to visualise or understand
what they are.

Flatpak allows apps to explicitly request permission to access
`/dev/kvm` and `/dev/shm`, so use the new flag for those.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #911