updates: Remove NO_DOWNLOAD flag from several GsPluginJobUpdateApps uses
We don’t actually care if the updates are downloaded at these call
sites, and in most cases the update will have already been downloaded
and cached so the runtime behaviour won’t actually change.

This improves support for plugins which don’t separate the download and
apply stages of their update job — currently it’s impossible for them to
work, as they can’t correctly put all of the update work into either the
download or the apply stage of the job.

Currently there aren’t any plugins like that, but the next few commits
will change the flatpak plugin so that it doesn’t separate download and
apply. This change is being made with the aim of making the flatpak
update process more robust against failure.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1929