flatpak: Stop splitting updates into download and apply stages
This is one step in reworking how the flatpak plugin updates apps. See
issue #1929 for the details.

This commit simply rearranges the code so that the `NO_DOWNLOAD` and
`NO_APPLY` flags are no longer checked for, and the two loops over
`applist_by_flatpaks` are now combined. That means that the code will do
all the downloads and applies on each `GsFlatpak` (a flatpak
installation) before moving on to the next one, rather than downloading
all the apps for each installation before applying them for each
installation.

The changes are simpler to visualise if this commit is viewed with `git
show -w` to ignore whitespace changes.

Subsequent commits will rearrange this code further.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1929