gs-repo-row: Expand translator comments for plural strings
As far as I know, the strings and the construction of the
multiple-plural string “%u apps and %u add-ons installed” is correct.

It was added in commit a8c4497697538, and discussed extensively in
https://bugzilla.gnome.org/show_bug.cgi?id=726990 beforehand. At least
fr, pt, es, sv and ar translations need the `cnt_apps + cnt_addon` count
to correctly translate “installed” to match the plurals in the sentence.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #2073