gs-details-page: Refresh buttons and progress when a job is added/removed
Use the new watch functions in `GsJobManager` to refresh the buttons and
progress for the details page’s app when a job related to that app is
added or removed from the job manager.

Eventually this will replace watching things like the app’s
`pending-action` property, since that’s limited to a single pending
action.

This change is needed now because it’s possible for the app’s state to
be updated by an ongoing `GsPluginJobUpdateApps` for the last time,
before the job completes. This means that the existing signal handlers
call `gs_details_page_app_has_pending_action()` for the last time while
it’s still going to return true (due to the not-yet-completed job).

Watching for job completion means we can update the UI again after the
job has completed.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1929