flatpak: Allow transactions to continue past the first fatal error
Currently, a `GsFlatpakTransaction` will stop on the first fatal error
it receives. Non-fatal errors are continued past.

This commit adds a property to allow fatal errors to be skipped, and
also changes the transaction to not stop on any non-fatal errors.

All the callers of `GsFlatpakTransaction` have been changed to continue
to use the existing behaviour of stopping on the first fatal error. Some
of them may be changed in future.

This commit mirrors the logic in the `flatpak` CLI. See
`stop_on_first_error` in `flatpak-cli-transaction.c`.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1929