eos-updater: Only try to update the app belonging to the updater
Currently, the `update_apps_async()` vfunc is called with a big list of
apps which are managed by all sorts of gnome-software plugins, not just
the plugin whose vfunc implementation is being called.

All the other plugins in gnome-software handle this correctly by
checking each app with `gs_app_has_management_plugin()` before trying to
update it.

I forgot to include those checks in the eos-updater plugin, so it was
assuming that it would only ever be called with the one `GsApp` for
eos-updater updates. That’s not the case, and led to an assertion
failure.

Fix that by filtering the app list down to the eos-updater app, which
might not be present.

In future, as per
https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2103, this could
change, but not this release cycle.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #2101