epiphany: Support D-Bus call flags for interactivity
This ensures that the service we’re communicating with can’t pop up an
unexpected polkit dialogue or other bit of UI if the gnome-software
operation is non-interactive.

This commit just changes the generated proxy API by adding the
`--glib-min-required` argument to `gdbus-codegen`. Default values are
used for all the calls, so there is no behavioural change.

The next commit will change the flags which are used. This should make
things easier to review.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>