gs-appstream: Always use first release description
When adding updates in gs_appstream_refine_app_updates(), there is unconditionally
used the first release's description when there's only one release available
in the appstream data, but when there are more versions, the first is skipped.
The app passed to this function is in an updatable state, thus the appstream
may contain the needed information. It might not be always accurate, because
the Flatpak apps can be released multiple times in the repository (different
commits), while having always the same 'release' information in the appstream
data of that app, but having there at least some information might look better
than "No description available".