rpm-ostree: Show updates download progress
The 'Download' button in the Updates page shows a progress of the app list
provided for the package download. Let the rpm-ostree plugin set the download
progress as reported by the rpm-ostree daemon, thus the users have a feedback
about the ongoing download.

There is emitted also the PercentProgress, but it's not for the download
itself, but for the respective operations being done within the update
process (which includes refresh of the rpm database for each configured
repository, each coming from 0 to 100, which can be confusing in the GUI).