flatpak: Change how app addons are pruned
The addons, the same as the apps itself, are cached and can be shared between
apps from different remotes within the same installation. As such, setting
"hide everywhere" on an addon can hide it not only for an app with which
the addon cannot be used, but also for other apps, where the addon is usable,
thus, rather than hiding the addons, remove them from the list of the addons
for the apps, where they cannot be used.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2152