fwupd: Report certain update failures to the user directly
`GsPluginJobUpdateApps` correctly does not report errors from plugins to
the user, as it can’t tell which plugin errors would be useful to the
user, and which are just transient or are noise.

`FWUPD_ERROR_NEEDS_USER_ACTION` is a fairly obvious candidate for being
reported to the user, though. It’s returned by fwupd when an update
failed for a reason which the user can potentially rectify.

This is similar to the user action handling code in
`gs_plugin_fwupd_install_async()`, except that it appears the
`FWUPD_ERROR_NEEDS_USER_ACTION` error code is only returned once an
operation has been tried, and it cannot be predicted before trying the
operation.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #2178