flatpak: Fallback to read ref info from <bundle> element when refining a wildcard
It can happen the appstream plugin constructs a new app from the provided component,
for example when refining a wildcard for an app, which is not cached by the flatpak
plugin yet. As the ref info is needed for many ops, make sure it is set also in such
cases and skip any apps, which do not have it set.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2176