packagekit: Report PkClientError errors in the GUI during package update
The package update can fail with various errors, but these are hidden
to the user, thus he/she cannot do anything about it. Show the errors
in the GUI to give the user a chance to know what went wrong.

It's not enough to pass the error to the caller, because the caller
only logs about it.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2189