fedora-third-party: Use plugin loader to determine which plugin handles `dnf` repositories
The fedora-third-party only provides repo types in a form of "flatpak" and "dnf",
which is not suitable for the Software, thus it had been hardcoded to use "packagekit"
as a replacement for the "dnf". This does not work on the Fedora Silverblue, because
it's the "rpm-ostree" plugin, which handles the "dnf" repositories, thus check which
plugin is enabled (using the plugin loader) and pick the correct replacement.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2208