gs-app: Drop a number of unused kudo values
Kudos as a concept have been dropped for a number of years now. Some of
them are still used internally in gnome-software as a cache for tracking
various positive attributes of apps. This is useful to avoid doing
AppStream queries when rendering the apps, and also to order apps when
returning search results.

The ones which are being removed in this commit are not set anywhere in
the code, so are completely dead.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>