shell: Don’t launch software-properties-gtk when opening repos dialog
This was perhaps a good idea 7 years ago when it was added, but since
then we’ve refreshed the repos dialogue, and it supports more different
types of packaging. In particular, `software-properties-gtk` doesn’t
support flatpak repos, but we want gnome-software to definitely support
that.

It appears `software-properties-gtk` was shipped in Debian and
(presumably) Ubuntu.

If they want to continue using that dialogue, they will have to patch
gnome-software to do so, or engage upstream to make sure the refreshed
repos dialogue from gnome-software covers their use cases.

This commit updates the draft release notes for 45.beta to make sure
this change is mentioned, as it’s relevant to packagers.

Spotted in #2231.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>