lib: Move rewrite-resources code out to the library
There’s no need for it to be inside the core
`gs-plugin-rewrite-resource.c` plugin. It probably won’t get reused
anywhere else, but by splitting it out to a separate file, it becomes
easier to unit test in future, and becomes possible to integrate it into
the core and drop the `gs-plugin-rewrite-resource.c` plugin.

Apart from making two functions public, this introduces no functional
changes.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>