appstream: Don’t use GError for control flow
Providing a GError to libxmlb results in the XbMachine internals
propagating errors to the caller beyond the basic G_IO_ERROR_NOT_FOUND.

This is not ideal because it means that a lot of CPU time is spent on
creating and freeing GErrors in tight loops that is otherwise
non-essential for control flow.

Another option would be to add API to libxmlb to denote the difference
between a result set, empty set (not found), and error but that will
require further coordination. See
https://github.com/hughsie/libxmlb/issues/149.