update-dialog: Port to AdwNavigationView
This is a rather big change, as it also involved porting GsOsUpdatePage
and GsAppDetailsPage to AdwNavigationPage. But overall it allowed
removeing a lot of code, because AdwNavigationView handles a lot of
things for us, like automatically adding a back button when required and
also removing pages on "pop" (when going back).

See
https://gnome.pages.gitlab.gnome.org/libadwaita/doc/main/migrating-to-breakpoints.html#navigation.