gs-download-utils: Cover case where `local_error` is set, but message status is 200
There can happen that the libsoup call fails, and sets the `local_error`, but
the message's status is left as 200, which is success. In such case
the g_assert() on `input_stream` being non-NULL triggers and aborts
the application.

(Backport to gnome-44 squashes commits 478afe452 and 71920b17f for
neatness.)

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2313