flatpak: Ignore read permission on xdg-config/kdeglobals
All KDE apps need this, and it’s harmless (it’s read-only access to
global data, nothing private), so allow it without ascribing additional
filesystem permissions to the app.

This means it will no longer show up in the app context bar as the app
having access to arbitrary files.

Signed-off-by: Philip Withnall <philip@tecnocode.co.uk>

Fixes: #2315