gs-details-page: Make spinner spin by default in "Loading app details..." page
This fixes the issue where the GtkSpinner is not displayed until the
details page switches to a non-default state (state !=
GS_DETAILS_PAGE_STATE_LOADING).

This is due to the 'state == gs_details_page_get_state (self)' check
in 'gs_details_page_set_state ()', which will be true for the first
few runs. Rather than handling this in code, we fix it in UI file.