lib: Add GS_APP_PERMISSIONS_FLAGS_SCREEN for screen contents access
This represents the other part of the data which access to Pipewire
(without using a portal) gives to apps, aside from that represented by
`GS_APP_PERMISSIONS_FLAGS_DEVICES`.

Add this to the UI and add an icon for it. Because this changes the UI
and translatable strings, it can only be added to a new major release of
gnome-software.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #2329