gs-updates-paused-banner: Inform about all updates paused conditions
Currently, gnome-software informs the user when automatic updates
are paused due to detecting a metered connection or Mogwai policy.
However, this isn't the only state that suspends automatic updates.

When in Power Saver mode, when GameMode is active and/or when large
downloads are not permitted by Mogwai, gnome-software won't
automatically update without informing the user as to why. Fix this
by teaching GsUpdatesPausedBanner to display information about these
other states.

This reintroduces a simple dialog that allows users to examine
the reason for paused updates when there are multiple.