packagekit: Show detailed error message when package install fails on GPG error
The PackageKit returns `PK_ERROR_ENUM_INVALID_PACKAGE_FILE` error when verification
of a package cannot be done, which used to be translated as "Not Supported" error
by the plugin, which does not show detailed error, even though the error message
shows explanation what failed. Translate these errors as "Failed" in the plugin,
to show the details to the user.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2343