CI: Create Docker image with local subprojects .wrap files
This does not download subprojects from the main branch on the gitlab.gnome.org,
but it uses the .wrap files from the local project (thus possibly with increased
versions), packs them together with Meson files into a zip file and extracts
it to cache the subprojects with them.

It makes it easier to update dependencies in the images.