ui: Ignore deprecated libadwaita API
The new version of libadwaita in the CI image has helpfully deprecated a
load of API without a trivial replacement. Rather than go through and
port the UI to use breakpoints in this merge request, disable warnings
about the deprecated UI, and wait for a proper fix from !1761, where the
UI is being ported to the newer version of libadwaita.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>
See: https://gitlab.gnome.org/GNOME/gnome-software/-/merge_requests/1761