CI: Switch to Fedora 39 in fedora.Dockerfile
This is to not build libadwaita, because it uses deprecated symbol
from appstream 0.16.4, which turns into a build break, due to the CI
build uses -Werror, which is not excluded from subprojects. As
the Fedora 39 has installed new-enough libadwaita the build of it
is not needed, the same as with the Debian image.

The exact build error follows:

```
../subprojects/libadwaita/src/adw-about-window.c: In function ‘adw_about_window_new_from_appdata’:
../subprojects/libadwaita/src/adw-about-window.c:2041:3: error: ‘as_component_get_developer_name’ is deprecated [-Werror=deprecated-declarations]
 2041 |   developer_name = as_component_get_developer_name (component);
      |   ^~~~~~~~~~~~~~
In file included from ../subprojects/appstream/src/appstream.h:28,
                 from ../subprojects/libadwaita/src/adw-about-window.c:9:
../subprojects/appstream/src/as-component.h:510:26: note: declared here
  510 | const gchar             *as_component_get_developer_name (AsComponent *cpt);
      |                          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
cc1: all warnings being treated as errors
```