rpm-ostree: Unregister client only when idle
Unregister client on the rpm-ostree daemon side only when the daemon
is idle, otherwise, when an ongoing operation takes longer than
the inactivity timeout, the client could be unregistered before
the transaction was finished.

The active transaction might not necessarily be from the gnome-software
itself, but it is not a big problem.