gs-appstream: Increase limit to category apps lookup
The Overview page currently checks for 100 apps in all categories.
Lookup for the same count per category.

Note: The counts are not accurate, same apps can be in multiple
sub-categories, thus the parent category count suffers of duplicity,
which the following commit will correct in a different way than
checking for id duplicity.