gs-update-monitor: Delay update time with larger random interval
It has been reported that some mirroring infrastructure sites
are suffering traffic spikes from update monitor from 6 to 7am.
This commit tries to resolve this issue by adding a randomized
hour parameter in [0, 6). And also, fallback to previous logic
if it has not been updated for 2 days or longer. GDateTime
methods are used to get datetime difference instead of handling
year, month and day values manually.

A more sophisticated strategy might be necessary if this problem
still exists after this commit, to both let users get updates
automatically, and not stressing infrastructure.

Fixes #2410.