fwupd: Change how historical updates are received
The fwupd plugin could error out when a returned update did not
have set appstream ID from the fwupd daemon, but it's a known state,
not an error, thus do not throw such failures to the user.

Also, read historical changes for each device separately, not only
for any device.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1163