gs-plugin-appstream: Reset state and invalidate silo when any plugin emits 'reload'
When any plugin emits a 'reload', the cached apps might not reflect the current
state, thus unset the state to have it refined when needed. Also invalidate
the silo, thus the installed apps hash is recreated.

The appstream plugin is special, because it serves as a helper for the other plugins.