packagekit: Print the running PackageKit version
This is to print the running packagekit version information in
gnome-software logs, in cases when packagekit has been upgraded /
installed manually and gnome-software was not rebuilt after that.

When running packagekit version (say, x.y.z) and the packagekit
version gnome-software was built with (say, a.b.c) are same, the
following message is printed.

PackageKit version: x.y.z

When the versions differ, the following message is printed.

PackageKit version: x.y.z (build version: a.b.c)