gs-os-update-page: Drop a new translatable string in favour of direction
Rework the previous commit to hard-code the two possible orders of
components in this string, based on the widget’s text direction. This is
the only way the string translation would have varied anyway.

This has the benefit of not breaking string freeze, and also making
things a bit simpler for translators.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #2437
Fixes: https://gitlab.gnome.org/Teams/Translation/Coordination/-/issues/142